split out logic from xfusuploader
